# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ProcessMaker API",
        "description": "",
        "contact": {
            "email": "info@processmaker.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/1.0"
        }
    ],
    "paths": {
...
        "/group_members": {
            "get": {
                "tags": [
                    "Group Members"
                ],
                "summary": "Returns all groups for a given member",
                "operationId": "getGroupMembers",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/member_id"
                    },
                    {
                        "$ref": "#/components/parameters/order_by"
                    },
                    {
                        "$ref": "#/components/parameters/order_direction"
                    },
                    {
                        "$ref": "#/components/parameters/per_page"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "list of group_members",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "description": "Display a listing of the resource.",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/groupMembers"
                                            }
                                        },
                                        "meta": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/metadata"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Group Members"
                ],
                "summary": "Save a new group member",
                "operationId": "createGroupMember",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/groupMembersEditable"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/createGroupMembers"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/group_members/{group_member_id}": {
            "get": {
                "tags": [
                    "Group Members"
                ],
                "summary": "Get single group member by ID",
                "operationId": "getGroupMemberById",
                "parameters": [
                    {
                        "name": "group_member_id",
                        "in": "path",
                        "description": "ID of group members to return",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully found the group members",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/getGroupMembersById"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Group Members"
                ],
                "summary": "Delete a group member",
                "operationId": "deleteGroupMember",
                "parameters": [
                    {
                        "name": "group_member_id",
                        "in": "path",
                        "description": "ID of group_members to return",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "success"
                    }
                }
            }
        },
        "/group_members_available": {
            "get": {
                "tags": [
                    "Group Members"
                ],
                "summary": "Returns all groups available for a given member",
                "operationId": "getGroupMembersAvailable",
                "parameters": [
                    {
                        "name": "member_id",
                        "in": "path",
                        "description": "ID of group member to return",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "member_type",
                        "in": "path",
                        "description": "type of group member to return",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "parameter": "filter",
                        "name": "filter",
                        "in": "query",
                        "description": "Filter results by string. Searches Name and Status. Status must match exactly. Others can be a substring.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "$ref": "#/components/parameters/filter"
                    },
                    {
                        "$ref": "#/components/parameters/order_by"
                    },
                    {
                        "$ref": "#/components/parameters/order_direction"
                    },
                    {
                        "$ref": "#/components/parameters/per_page"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "list of groups available to be assigned as member",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "description": "Display a listing of groups available",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/availableGroupMembers"
                                            }
                                        },
                                        "meta": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/metadata"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user_members_available": {
            "get": {
                "tags": [
                    "Group Members"
                ],
                "summary": "Returns all users available for a given group",
                "operationId": "getUserMembersAvailable",
                "parameters": [
                    {
                        "name": "group_id",
                        "in": "path",
                        "description": "ID of group to return",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "parameter": "filter",
                        "name": "filter",
                        "in": "query",
                        "description": "Filter results by string. Searches Name. Can be a substring.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "$ref": "#/components/parameters/order_by"
                    },
                    {
                        "$ref": "#/components/parameters/order_direction"
                    },
                    {
                        "$ref": "#/components/parameters/per_page"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "list of users available to be assigned as member",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "description": "Display a listing of users available",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/users"
                                            }
                                        },
                                        "meta": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/metadata"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "groupMembersEditable": {
                "properties": {
                    "group_id": {
                        "description": "Represents a group Members definition.",
                        "type": "string",
                        "format": "id"
                    },
                    "member_id": {
                        "description": "Represents a group Members definition.",
                        "type": "string",
                        "format": "id"
                    },
                    "member_type": {
                        "description": "Represents a group Members definition.",
                        "type": "string"
                    },
                    "description": {
                        "description": "Represents a group Members definition.",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "groupMembers": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/groupMembersEditable"
                    },
                    {
                        "properties": {
                            "id": {
                                "description": "Represents a group Members definition.",
                                "type": "string",
                                "format": "id"
                            },
                            "created_at": {
                                "description": "Represents a group Members definition.",
                                "type": "string",
                                "format": "date-time"
                            },
                            "updated_at": {
                                "description": "Represents a group Members definition.",
                                "type": "string",
                                "format": "date-time"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "createGroupMembers": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/groupMembersEditable"
                    },
                    {
                        "properties": {
                            "id": {
                                "description": "Represents a group Members definition.",
                                "type": "string",
                                "format": "id"
                            },
                            "group": {
                                "description": "Represents a group Members definition.",
                                "type": "object"
                            },
                            "member": {
                                "description": "Represents a group Members definition.",
                                "type": "object"
                            },
                            "created_at": {
                                "description": "Represents a group Members definition.",
                                "type": "string",
                                "format": "date-time"
                            },
                            "updated_at": {
                                "description": "Represents a group Members definition.",
                                "type": "string",
                                "format": "date-time"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "getGroupMembersById": {
                "allOf": [
                    {
                        "properties": {
                            "group_id": {
                                "description": "Represents a group Members definition.",
                                "type": "string",
                                "format": "id"
                            },
                            "member_id": {
                                "description": "Represents a group Members definition.",
                                "type": "string",
                                "format": "id"
                            },
                            "member_type": {
                                "description": "Represents a group Members definition.",
                                "type": "string"
                            },
                            "id": {
                                "description": "Represents a group Members definition.",
                                "type": "string",
                                "format": "id"
                            },
                            "created_at": {
                                "description": "Represents a group Members definition.",
                                "type": "string",
                                "format": "date-time"
                            },
                            "updated_at": {
                                "description": "Represents a group Members definition.",
                                "type": "string",
                                "format": "date-time"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "availableGroupMembers": {
                "allOf": [
                    {
                        "properties": {
                            "id": {
                                "description": "Represents a group Members definition.",
                                "type": "string",
                                "format": "id"
                            },
                            "description": {
                                "description": "Represents a group Members definition.",
                                "type": "string"
                            },
                            "name": {
                                "description": "Represents a group Members definition.",
                                "type": "string"
                            },
                            "status": {
                                "description": "Represents a group Members definition.",
                                "type": "string",
                                "enum": [
                                    "ACTIVE",
                                    "INACTIVE"
                                ]
                            },
                            "created_at": {
                                "description": "Represents a group Members definition.",
                                "type": "string",
                                "format": "date-time"
                            },
                            "updated_at": {
                                "description": "Represents a group Members definition.",
                                "type": "string",
                                "format": "date-time"
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
            "metadata": {
                "properties": {
                    "filter": {
                        "type": "string"
                    },
                    "sort_by": {
                        "type": "string"
                    },
                    "sort_order": {
                        "type": "string",
                        "enum": [
                            "asc",
                            "desc"
                        ]
                    },
                    "count": {
                        "type": "integer"
                    },
                    "total_pages": {
                        "type": "integer"
                    },
                    "current_page": {
                        "type": "integer"
                    },
                    "form": {
                        "type": "integer"
                    },
                    "last_page": {
                        "type": "integer"
                    },
                    "path": {
                        "type": "string"
                    },
                    "per_page": {
                        "type": "integer"
                    },
                    "to": {
                        "type": "integer"
                    },
                    "total": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        },
        "parameters": {
            "filter": {
                "name": "filter",
                "in": "query",
                "description": "Filter results by string. Searches Name, Description, and Status. Status must match exactly. Others can be a substring.",
                "schema": {
                    "type": "string"
                }
            },
            "order_by": {
                "name": "order_by",
                "in": "query",
                "description": "Field to order results by",
                "schema": {
                    "type": "string"
                }
            },
...
            "order_direction": {
                "name": "order_direction",
                "in": "query",
                "schema": {
                    "type": "string",
                    "default": "asc",
                    "enum": [
                        "asc",
                        "desc"
                    ]
                }
            },
            "per_page": {
                "name": "per_page",
                "in": "query",
                "schema": {
                    "type": "integer",
                    "default": "10"
                }
            },
...
            "member_id": {
                "name": "member_id",
                "in": "query",
                "schema": {
                    "type": "integer"
                }
            },
...
        },
        "securitySchemes": {
            "pm_api_auth_code": {
                "type": "oauth2",
                "flows": {
                    "authorizationCode": {
                        "authorizationUrl": "/oauth/authorize",
                        "tokenUrl": "/oauth/token",
                        "scopes": "*"
                    }
                }
            },
            "pm_api_bearer": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            },
            "pm_api_key": {
                "type": "apiKey",
                "in": "header",
                "name": "Authorization"
            }
        }
    },
    "security": [
        {
            "pm_api_token": []
        },
        {
            "pm_api_key": []
        },
        {
            "pm_api_auth_code": []
        }
    ]
}